+++ modules/laptop.nix
@@ -1,0 +1,10 @@
+{
+  hmModule = { ... }: {
+    dconf.settings = {
+      "org/gnome/desktop/peripherals/touchpad" = {
+        natural-scroll = false;
+        tap-to-click = true;
+      };
+    };
+  };
+}